refactor: reorganize AI functions with proper directory structure
- Refactor ai-abstract-generator and ai-tag-generator to follow directory convention
- Separate types, validators, and services into dedicated files
- Add proper admin authentication with JWT metadata checking
- Improve code organization and maintainability
- Abstract service handles mock abstract generation with style options
- Tag service manages fixed tags and confidence scoring
- Input validation now follows consistent pattern across all functions
